miss2/root/requirements.txt
dnspython==2.0.0
websocket-client==0.57.0